backend/requirements.txt
fastapi
uvicorn
requests
sentence-transformers
faiss-cpu
torch
transformers
peft
bitsandbytes
